benchmark_app/requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
azure-identity>=1.17.0
azure-storage-blob>=12.20.0
azure-ai-documentintelligence>=1.0.0
openai>=1.30.0
